requirements.txt
streamlit
scikit-learn
pandas
faiss-cpu
sentence-transformers
openai
python-dotenv
